requirements.txt
# Core dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Data processing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
mne>=1.5.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
